feat: implement calculated metrics (metrics which can reference columns/dimensions/metrics from multiple models)
Implement `aggregated()` function in MetricAliasMetricRef class. When this function is called,
referenced metric will be sliced only across the explicitly queried dimensions. Otherwise primary
key dimensions are implicitly added to the outer query which is useful if you want to perform
additional aggregation on aggregated data.
Implement `groupBy` function in MetricAliasColumnOrDimensionRef class. When this function is called
the referenced column or dimension will participate in the GROUP BY on segment level.